requirements.txt
langchain>=0.3
langgraph>=0.2
langchain-ollama
langchain-google-genai
streamlit
mlflow
python-dotenv
faiss-cpu
jupyter
langchain-community 
langchain-core
chromadb
pypdf